Added check that only the creator of the quiz can delete it

+++ app/src/main/java/comp3350/quizrus/business/AccessQuizzes.java
@@ -48,7 +48,15 @@ public class AccessQuizzes {
         return newQuiz;
     }
 
-    public void deleteQuiz(Quiz quiz) {
-        quizPersistence.deleteQuiz(quiz);
+    public boolean deleteQuiz(Quiz quiz, User user) {
+        if(quiz.getUser().getUserID() == user.getUserID())
+        {
+            quizPersistence.deleteQuiz(quiz);
+            return true;
+        }
+        else
+        {
+            return false;
+        }
     }
 }
